% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{export_maze_3d}
\alias{export_maze_3d}
\title{Export text coordinates of Maze object walls}
\usage{
export_maze_3d(maze, filename)
}
\arguments{
\item{maze}{a Maze object}

\item{filename}{name of output .txt file}
}
\description{
Export text coordinates of Maze object walls
}
