% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{render_maze_tri}
\alias{render_maze_tri}
\title{Render a Triangular Grid Maze object}
\usage{
render_maze_tri(maze, show_solution = FALSE)
}
\arguments{
\item{maze}{a Maze object}

\item{show_solution}{a logical describing whether to plot the solution to the maze}
}
\value{
ggplot of Maze object
}
\description{
Render a Triangular Grid Maze object
}
\examples{
generate_maze_tri()
}
