% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{render_maze}
\alias{render_maze}
\title{Render a Maze object}
\usage{
render_maze(maze, show_solution = FALSE)
}
\arguments{
\item{maze}{a Maze object}

\item{show_solution}{a logical describing whether to plot the solution to the maze}
}
\value{
ggplot of Maze object
}
\description{
Render a Maze object
}
\examples{
library(tidyverse)
generate_maze()
}
